requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.19.0
altair==5.3.0
annotated-types==0.7.0
anthropic==0.29.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bitarray==2.7.3
bitsandbytes==0.41.0
bleach==6.0.0
cattrs==22.2.0
cdsw @ file:///./runtime/engine/python-module
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
cmlapi==23.11.28
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.18.0
debugpy==1.6.7
decorator==4.4.2
deepspeed==0.14.2
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring-to-markdown==0.12
e==1.4.5
einops==0.6.1
einops-exts==0.0.4
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.16.3
ffmpy==0.3.2
filelock==3.14.0
flash-attn==1.0.2
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
gitdb==4.0.10
GitPython==3.1.37
google-pasta==0.2.0
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.23.0
idna==3.4
importlib_resources==6.4.0
impyla==0.18.0
ipykernel==6.22.0
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
jedi-language-server==0.40.0
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==7.2.2
jupyter-events==0.6.3
jupyter-lsp==2.1.0
jupyter-server==1.23.4
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.4
jupyter_core==5.3.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.0
jupyterlab-git==0.41.0
jupyterlab-lsp==4.1.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keras==3.4.1
kerberos==1.3.1
kiwisolver==1.4.4
kollava==1.1.3
libclang==18.1.1
linkify-it-py==2.0.3
lit==18.1.4
llava @ file:///home/cdsw/KoLLaVA
lsprotocol==2023.0.0a1
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.3.2
mlflow-CML-plugin==0.0.1
mlflow-skinny==1.19.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
nbclassic==0.5.3
nbclient==0.6.3
nbconvert==7.2.9
nbdime==3.2.0
nbformat==5.7.3
nest-asyncio==1.5.5
networkx==3.2.1
ninja==1.11.1.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
peft==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.5.0
preprocess==2.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.6.0
pure-eval==0.2.2
pure-sasl==0.6.2
py-cpuinfo==9.0.0
py4j==0.10.9.1
pycparser==2.21
pydantic==1.10.17
pydantic_core==2.20.1
pydub==0.25.1
pygls==1.0.1
Pygments==2.15.1
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0
pyzmq==22.3.0
regex==2024.4.28
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
ruff==0.5.0
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==2.7.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.5.0
thrift==0.16.0
thrift-sasl==0.4.3
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchaudio==2.1.2+cu121
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
uc-micro-py==1.0.3
ujson==5.9.0
uri-template==1.2.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.17.4
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.16.0
y-py==0.5.9
yarl==1.9.4
ypy-websocket==0.8.2
